キャッシュモジュールを作ってそこから取得してみる

+++ src/sample/controller/test.go
@@ -101,7 +101,7 @@ func Test(c *gin.Context) {
 
 	// COUNT TEST
 	condition := model.Condition{"where": model.WhereCondition{{"id", "IN", model.In{1, 2, 3}}}}
-	option = model.Option{"shard_key": 1}
+	option = model.Option{"shard_key": uint64(1)}
 	count, err := userRepo.Count(c, condition, option)
 	if checkErr(c, err, " count error") {
 		return
